FloatingActionButton
It has three overloads:
1) With Custom content: useful for something custom like animated icon transitions
2) With an Image: the most common case
3) Extended FAB: Has text and an icon.

Specifications: https://material.io/design/components/buttons-floating-action-button.html
I decided to not implement the Mini Fab for now. Looks like the main use case for it is to display them instead of the default one if the screen size is too small. It means our default FAB should be smart enough to understand it and change the default size instead of providing the separate MiniFloatingActionButton component. We will figure out how it can be implemented later. For other use cases it will be anyway really easy to provide a custom size for the FAB.
I also had to fix Spacer component as I found it doesn't work nicely with the cases when we want to put Spacer into Wrap container, it will always fill all the parent size so it breaks the wrapping. Now it will have a zero sized secondary side.

Bug: 113858259
Test: new demo screen, tests for correct layout sizes, later will be worth adding screenshot tests
Change-Id: I6697172b52621b7f1de4092967c11402eb418393