Discard retained Fragments not in the active set on restore
In very specific instances, an Activity can have its
state saved, be stopped, then be started again. This
puts FragmentManager in a state where it thinks the
state is not saved, but it actually is still saved
(and the Activity will subsequently be stopped, destroyed,
then recreated with the old saved instance state).

Rather than keep the state as not saved even
though the Activity is started (which would break
some of the guarantees we provide for safety),
we'll instead discard any retained Fragments that were
added in that brief window of started but unsaved
so that we'll be back to the consistent state found
in the saved instance state.

Test: updated SaveStateFragmentTest
Fixes: 130289043
Change-Id: I3fe1a788ef3f3ace0ff18ba405a15b63708e41f7