Change save fragment max lifecycle test save STARTED
StrictFragment requires that the field variable 'currentState' be
equal to CREATED when onDestroy is called or it will throw an
IllegalStateException. This causes problems with testing with
MaxLifecycleState.

When using StrictFragment, if MaxLifecycle is set to CREATED for a
Fragment and the FragmentController is restarted (or shutdown), the
Fragment will not dispatchPause or dispatchStop, because it remains
in the ACTIVITY_CREATED state. This means that 'currentState' is equal
to ACTIVITY_CREATED when onDestroy is called and the exception is
thrown.

This test was passing because setMaxLifecycle was split into two Ops
(set and unset max lifecycle) and as a result the Fragment was being
forced to the RESUMED state, then going through the proper transitions
to onDestroy().

Test: ran this test
Change-Id: I7382b3f712b0e03480b6f8708e3a18a89c0a52b3