Remove camera2InteropCaptureSessionCallbacks() temporary
Because the implementation of burst capture, the capture result changed
to CONTROL_CAPTURE_INTENT_PREVIEW.
Remove camera2InteropCaptureSessionCallbacks() temporary,

And this test case will be considered to rewrite or review and also tracking in another bug.

If there is any patch about captue request or burst capture, this test
case should be review again.

Bug: 131336029, 130817150
Test: manual test with emulator
Change-Id: Iadc5ca8edd6196c6032438fbef2bbdedd1e7a4d0