Fix OnChildPositioned to work with multiple children in the same parent
It was working incorrectly as was just having one list of the listeners in parent and if this parent will have two child each wrapped into OnChildPositioned both of the callbacks would be called with the positions of both children. New implementation is using ambients to provide the list of the current OnChildPositioned callbacks into the children instead, so it affects only the children within the lambda of the OnChildPositioned component.

Test: new tests added
Change-Id: Id5cc008829fcb8867bb89c83524718a67a4dc027