Check that Fragments are still hidden before animating
If a fragment is hidden with an animation and set to be shown again
before the animation completes, the animation will be executed later,
causeing the fragment to disappear. We do not have a reference to the
animator after it sets the AnimatorListenerAdapter and as a result, we
cannot cancel the animation.

This change checks that the Fragment should still be hidden before
setting its visibility to View.GONE.

Test: Added Test, tested in app, all test pass
BUG: 133385058
Change-Id: I400b71dcb3f1c64e0cf9e5a2a9d9a82a2b10dfd6