Ensure ActionBar is not updated when destination is floating window
When showing a dialog destination, the actionbar updates to show an up
arrow. This behavior is unexpected and it is not possible to use the
back button. The action bar should remain the same when interacting with
any floating window.

This change adds a marker inferface, FloatingWindow, that should be
implemented by classes considered to be floating windows.
AbstractAppBarOnDestinationChangeListener checks if the destination is an
instance of FloatingWindow to determine if the AppBar should be
updated.

Added label for dialog destination in integration app for testing.

Test: Tested in integration app. ./gradlew checkApi
BUG: 133600763
Change-Id: Ie4efab6c265eca13583bd21615f8d571c2236443