Allow using POJOs in @Insert, @Update and @Delete method parameters.
Add a new method to @Insert, @Update and @Delete called 'entity', that
allows declaring the target entity the method is meant to affect. The
method will interpret the POJO parameters as partial entities.

If the method is an INSERT, we allow any POJO with a subset of columns
that will not cause a NOT NULL constraint. If the real entity has a
auto generated primary key, then that column is optional too.

If the method is an UPDATE or DELETE then we allow any POJO with a
subset of columns regardless of the schema default values. For UPDATE
the only requirement is that the POJO must have fields matching the
primary keys of the entity.

Bug: 127549506
Test: ./gradlew room:room-compiler:test \
                room:integration-tests:room-testapp:cC
Change-Id: Ie0477d91bb92f799486ff672b20d4564769903d2