Remove quality setting from CameraView
Removed setQuality API and internal references to quality.  The
resulting quality will just be the default CameraX quality settings.

The existing controls were not affecting current behaviour, and the
settings are not finalized, and only HIGH was acceptedd as a setting.

Clearer to remove this for now, then re-add as a full feature
(see Bug ID: 123894197)

Bug: 126626274
Test: ./gradlew :camera:integration-tests:camera-testapp-view:build and
manual testing.

Change-Id: I7407777b5ab9413f2bef77d58a4e634099aff515