Rename CameraView.setCameraByLensFacing
Rename CameraView.setCameraByLensFacing to
CameraView.setCameraLensFacing
Makes CameraLensFacing set/get work a s property in Kotlin.
While setCameraByLensFacing is a bit clearer, having this work as
a property is better and there's no real functionality to 'set
the facing of the camera' so should be minimal confusion.
Also renamed in the internal CameraXModule to match.

Bug: 128230637
Test: ./gradlew :camera:integration-tests:camera-testapp-view:installDebug and manual test
Change-Id: I55dc29ed5b190d59566c63745f72894dc2d1ef8e