Register listeners before scrolling
CompositeOnPageChangeCallback is not thread safe, so registering
listeners while scrolling can lead to race conditions. In postsubmit
this led to a tool failure with this stacktrace:

java.lang.NullPointerException: EMPTY
java.lang.NullPointerException
        at androidx.viewpager2.widget.CompositeOnPageChangeCallback.onPageScrolled(CompositeOnPageChangeCallback.java:59)
        at androidx.viewpager2.widget.CompositeOnPageChangeCallback.onPageScrolled(CompositeOnPageChangeCallback.java:59)
        at androidx.viewpager2.widget.ScrollEventAdapter.dispatchScrolled(ScrollEventAdapter.java:410)
        at androidx.viewpager2.widget.ScrollEventAdapter.onScrolled(ScrollEventAdapter.java:188)
        at androidx.recyclerview.widget.RecyclerView.dispatchOnScrolled(RecyclerView.java:5152)
        ... and 42 more

Did 29 runs of 2000 iterations before this fix to find the probability
of the tool failure, which is estimated to be 0.002634
Did 11 runs of 2000 iterations to verify that the tool failure doesn't
occur anymore. Assuming a geometric distribution, the probability that
this CL does _not_ fix the tool failure is 6.252E-26.

Bug: None
Test: 11x 2000x ./gradlew viewpager2:connectedCheck \
        -Pandroid.testInstrumentationRunnerArguments.class=\
        androidx.viewpager2.widget.PageChangeCallbackTest

Change-Id: I5c803809e74c7925d5e83f19a803a0de86c8e9ef