Remove setMaterialContentAndCollectDpSize test api
This method is dangerous as will return not the results we expect.
Example is asserting height of the button is 36.dp.
If we use:
setMaterialContentAndTestSizes { Button() }
   .assertHeightEqualsTo(36.dp)
It works nice, but if we try to use removed method:
val size: Size = setMaterialContentAndCollectDpSize { Button() }
assertThat(size.height).isEquals(36.dp)
this will fail as size.height is 36.2 dp(depends on the density)
The reason is we round float pixels to IntPx for layouting, but this
method is converting rounded ints back to Dp which would never return
us expected value of 36.dp.
We should always compare pixel sizes instead so this method is removed.

Test: nothing, this test util method was unused
Change-Id: Icd4d7c27a1c9d81de320ab7ef82a40fd5caa024c