Ensure postponedEnter with duration only starts once
Right now if the postponedEnterTransition(duration, TimeUnit) is called
twice, startPostponedEnterTransition is also called twice. The proper
behavior is for the first Runnable to be replaced if a second is called
before the duration time runs out.

This change maintains the current Runnable in the Fragment's
AnimationInfo and if postponedEnterTransition is called while another
runnable already exists that Runnable is removed from the handler.

Test: Added Test
BUG: 137797118
Change-Id: Iecd55a1267e7613d025e0b95393996ad53dcace3