Introduce LayoutResult for layout() layoutResult()
To make sure that layout(..) and layoutResult(..) are called from the
layout lamba of Layout and ComplexLayout, respectively, the lambdas are
now required to return a LayoutResult object. This object can only be
obtained by calling layout(..) or layoutResult(..), which guarantees
that those methods will be called from those lambdas.

Bug: 135941724
Test: ui:ui-framework:cC

Change-Id: Ic90cb7f2cce401128171e929cf8827a37bb9f7e7