Add alignment lines for text first&last baselines
The CL defines alignment lines corresponding to the first and last
baseline of Text widgets. These can be used by parent layouts to
position children according to baseline. Also, the baselines are
inherited by parents - if two children of the same layout provide
baselines, the parent will choose the minimum one for FirstBaseline and
the maximum one for LastBaseline.

Bug: 129859180
Test: ran TextLayoutTest#testTextLayout_providesBaselines
Change-Id: I6d71907704880c13b1b1623cc83fba7b27d6da42