Conditionally error due to missing getter/setter based on binding scope.
When processing a POJO with binding scope to bind to query only, then
it is OK if the POJO has no setter or constructor. Likewise if a POJO
with binding scope to read from cursor, then it is OK to have no
getter. When the binding scope is two-way, then both getters and
setters are needed (or getter and matching constructor).

Bug: 138664463
Test: PojoProcessorTest
Change-Id: I1324f2e55174aa9999498666000b86591d05e970