Add BiometricPrompt device credential support for L+
Currently, BiometricPrompt only supports authentication with device
credentials (e.g., PIN, pattern, or password) on Android Q. This is
because prior versions implemented device credential authentication by
passing an activity result. Because BiometricPrompt is hosted in a
client activity, this would cause the client activity to receive the
result, leading to incorrect behavior.

This commit builds on the previous one that introduced support for
device credential authentication to P by also adding for versions L to O
(which use androidx.biometric.FingerprintDialogFragment internally). To
do so, it builds on top of the existing handler activity solution while
adding some additional fixes and customizations for the fingerprint
fragments.

Test: Biometric support demo app on Android M, P, and Q devices

Bug: 123595980
Change-Id: I8b70964b256d1315538b9ce3ff85d54332acea9d