Allow developers to write their own column width specs
This allow developers to implement their own instances of
TableColumnWidth by inheriting from its abstract subclass
TableColumnWidth.Inflexible. This has an abstract function called
preferredWidth() which computes the ideal width of that column and
replaces the 'when' expression in the existing definition of Table.

Furthermore, this introduces a method flexible(flex) that creates
a column width specification that will default to the width given
by preferredWidth(), but may grow to fill any additional space left.

Bug: 137850753
Bug: 138295535
Test: Ran TableTest and demo in TableActivity
Test: Added testTable_withColumnWidth_[wrap_flexible/custom]
Change-Id: I835421779bd8b3c2f4bf7fa40b78c5bab2f98074