Simplify usage of RoundedCornerShape/CutCornerShape
Provide more overloads for RoundedCornerShape CutCornerShape so we don't have to explicitly create CornerSizes class. There is no real need in CornerSizes anymore so we remove this class together with PxCornerSizes.
This should simplify the usages and improve readability. New methods are intended to be marked with inline when this would be supported.

Bug: 139271525
Test: RoundedCornerShapeTest, CutCornerShapeTest passed
Change-Id: I062c83bf5b8493fccc5e7cdd0d97fd658a92bccb