Rework Scroller to make it flingable
This CL adds fling support to Scroller via providing fling when touhc
input has ended and by providing smoothScrollTo method in
ScrollerPosition. This change requires animations to be in Scroller,
therefore it's moved from ui-layout to ui-foundation.

The addition of fling required serious rework to make it as performant
as it was but with animations support. First, DragGestureDetector was
replaced with Draggable, which supports various controllers and
animations support by one of these controllers. This Draggable now is
top level as we can't really afford to move GestureDetector together
with the layout (both of velocity and performance).

To stop in-progress fling by touch, one more InputGestureDetector was
added. Also, demo for Horizontal controlled scrolles is now here with
buttons to demonstrate smooth scrolling and case when isScrollable ==
false.

Added some documentation to better explain params in Scroller as well as
Samples.

Bug: 139470709
Test: ./gradlew ui:ui-foundation:connectedCheck passed with ScrollerTest
moved from layout to foundation module
Change-Id: I92301ede2fb2c7c85441c656a1756952b8321616