NavController: disallow setting ViewmodelStore after setGraph
Otherwise:
1. NavBackStackEntry can be in CREATED state but ViewModels aren't
accessible even though they are supported by NavBackStackEntries
2. One instance of ViewModelStore can be replaced with another one
in NavBackStackEntry while it is alive. It breaks invariant that
if you do viewmodelProvider.get(Foo.class) twice on the same instance
of ViewModelStore you'll get the same instance.

In general the safe pattern is to call setViewModelStore before using
NavController (like we do in NavHostFragment) ans we should force
devs to do the same to avoid footguns.

Test: ./gradlew -p navigation connectedCheck
Change-Id: Ib83fc4387ea3cb70d2125eb1ada9be2425ef3a94