Use Log.isLoggable() for FragmentManager debug logging
Instead of requiring developers rebuild their app
with FragmentManager.enableDebugLogging(true), we
should also respect Log.isLoggable() to determine
what VERBOSE and DEBUG logging to output.

This also updates logging tied to moveToState()
to use the DEBUG level, ensuring that this critical
logging is more noticable when logging is enabled.

Test: ./gradlew checkApi
Change-Id: Ia3a0eddac00ef2a159f1848959b6a5a578a44e76