Let BiometricPrompt directly launch CDC activity
Currently, BiometricPrompt fails with an error when attempting to
authenticate with no supported biometrics enrolled on API < 29. If
allowing device credential auth, however, the prompt could simply fall
back to that form of authentication in this case. This commit makes that
change, bringing support for this option up to parity with API 29.

Test: ./gradlew biometric:connectedAndroidTest
Test: Manually, using biometric demo app on API 23, 27, 28, and 29.

Bug: 140750912
Change-Id: I863daed71c1267f0670ba8dfc91d0c95bf46cd2c