Fix issue where Camera cannot be reopened after disconnect
Handle the onDisconnected() callback in CameraDevice so that the Camera
can gracefully reopen or close. The CaptureSession now has a
forceClose() method that shortcircuits the
CameraCaptureSession.StateCallback.onClose() from the framework and
instead calls it directly because bug (b/140955560) prevents the it from
being called successfully.

Treat the disconnected callback as equivalent to an error state where
the camera is in use

Removed the testDisconnect_launchNativeCameraApp test because it is not
controlled by CameraX.

Fix issue in ImageAnalysisTest where mAnalysisResult could be accessed
by multiple threads

Bug: b/140631385, b/139498531, b/139904159, b/139261219
Test: ./gradlew camera:camera-camera2:connectedCheck

Change-Id: Iddc862952ae375855ef2babb7178c791b0e6c928