Fix broken AccessibilityTest
performAccessibilityAction should be called on the UI thread. For some
reason it never surfaced before, presumably because the scrolling that
it triggers was performed in onAnimation and not as a direct result of
the action. Now we clear focus, however, this action triggers a change
in the view and the test crashes.

Bug: 140656866
Test: ./gradlew recyclerview:recyclerview:cC \
        -Pandroid.testInstrumentationRunnerArguments.class=\
        androidx.viewpager2.widget.AccessibilityTest

Change-Id: Ia7113fac1f9152ddaea4499a56220fea09822407