NotificationCompat: Fix null-pointer exception on API level 23
This is a fix for a bug caused by
https://android.googlesource.com/platform/frameworks/support/+/db66df0a6b8bc0f5a0c7439189c69be9100b7421

In Android M (23) we started using Icon instances instead of just
resource IDs to represent icons in Notification Actions. But
NotificationCompat#getActionCompatFromAction() still (before this CL)
uses the old resource ID. This is a problem on API level 23 (not before
or after that) since on API level 23 we don't copy the resource ID from
the Icon object (Action.mIcon) into the old resource ID (Action.icon),
thus causing us to lose information about the Icon in
NotificationCompat#getActionCompatFromAction().

With this CL we use the new Icon object in
NotificationCompat#getActionCompatFromAction() on API levels 23 and above.

Bug: 141457324
Test: NotificationCompatTest
Change-Id: Ide3b7df3997317223dd7f7c3aed36fe73201b182