Enable trackRestrictAPIs for media2-session
This CL adds empty restricted_1.0.0-rc01.txt in order to successfully
generate restricted_current.txt as it seems that updateApi tries to
read the latest restricted txt file. restricted_1.0.0-rc01.txt could
be re-generated based on the snapshot but there unintentionally were
some broad @RestrictTo usages in the module, so 1.1.0 could be
considered to be the first version that tracks restricted apis
correctly.

Fix: 131782284
Test: ./gradlew -p media2 check
Change-Id: I57c3f37cd144f79efd0052ec6727fc6d7ea82e25