Remove RippleSurface, simplify Ripple implementation
Originally we introduced the approach similar to what we have in Android Framework and Flutter where some top-level surface emits RippleSurface and we draw RippleEffects on this level, not on the level where the clicked layout was added. Seems like with the flexibility of Compose and our decision not to have clip to bounds by default we can now simplify this. Ripple can instead emit Draw node before the target layout(in it's parent), and then we can just translate to the local target clayout position. This behaviour can even be more predictable then before, as previously if you emit Surface with full sized image and add button on top of the image Ripple would be drawn below the image. With the current approach we will naturally draw on top of the image instead.

This also simplifies Surface as now Surface has no relation to Ripples, so I cleared the documentation for Surface, Card and TransparentSurface and in the end decided to remove TransparentSurface at all as seems like we don't really encourage people to use it anymore. Previously you were able to use it just to provide a surface for inner ripples, but now they would work without a surface.

This refactoring allowed to simplify the logic we have in RippleEffect:
1) We can remove the logic of converting coordinates from the surface to the target layout
2) We can remove Matrix transformation as seems like we would never need it as we can apply them directly on Canvas instead for things like rotation.
3) We can resolve some params right in Ripple components instead of passing all the information to DefaultRippleEffect just to figure out that the effect should start from center, not from the touch position
4) We can provide color as a param for RippleEffect.draw() instead of updating the field

Bug: 141676126
Test: tested manually. two Ripple tests currently don't work as they were relaying on the removed mechanism of matrix translation. In the follow up CL I am going to move testing bitmap capturing logic from ui-framework to ui-test so we can reuse it here and rewrite them to assert what was really drawn instead of comparing matrix.
Change-Id: Ifd1501d0f227fd7c7b54caa918a9f326385b1633