Add enabled param for Ripple
If Ripple is disabled it will still consume the correlated pointer events, but the effect wouldn't be started.
It allows to simplify the usage of Ripple in Buttons. Instead of:
f (onClick != null) {
    Ripple(bounded = true) {
       Clickable(onClick = onClick, children = children)
    }
} else {
    Clickable(children = children)
    }
}
We can now write just:
Ripple(bounded = true, enabled = onClick != null) {
    Clickable(onClick = onClick, children = children)
}
It is also more correct semantically, as Clickable is still consuming the pointer events even when disabled, but Ripple wasn't following the same pattern previously.

Bug: 141972529
Test: added new test
Change-Id: I6158403af879d00e1940839bfb8abf2716c26e21