Remove BaseButton
With the recent change for Ripples they not require Surface anymore.
So now if you want to have a clickable content with Ripple instead of using BaseButton you can just write:
Ripple(bounded = true {
   Clickable(onClick = onClick) { ... }
}
So the abstraction of BaseButton isn't really needed anymore:
a) the implementation is already simple enough
b) currently all the params from Ripple should be duplicated to params of BaseButton
c) If you will want to have unbounded ripple with BaseButton you can't do so because BaseButton also wraps the content into Surface, and Surface will clip the ripple. It will be inconsistent to use BaseButton for bounded clickable areas, and something else for unbounded.

Bug: 141972529
Test: manually
Change-Id: I77c69790d31bd0dc7bc556fe6908a1baefe76cf2