Only save original delegates in if getItemDelegate() returns an ItemDelegate
PreferenceRecyclerViewAccessibilityDelegates item delegate calls into the original implemtnation of itemDelegate
through composition not inheritance. Since we were saving original item delegates on mItemDelegate,
but setting the delegate from getItemDelegate, when a view was bound twice, without being sent to the pool,
a call to an accessibility function would result in an infinite loop between getItemDelegate() and mItemDelegate.

Test: RecyclerViewAccessibility* tests.
Fix: 142279381

Change-Id: Id4ed6f76ce90a767f68707bd2cc26256fb62f097