Prevent AnnotatedString build changing internal state
AnnotatedString.Builder#build function would pop from the stack which
caused the internal state to change. This would cause issues if build
was called in lambdas such as withStyle. This CL changes that behavior.
Now build() does not change internal state.

Moreover in order to comply with kotlin style toXYZ and also StringBuilder,
changed the function name to be toAnnotatedString.

Test: ./gradlew ui:ui-text:test
Test: ./gradlew ui:ui-text:connectedAndroidTest

Bug: 139312718
Change-Id: I3ec66ea0c77b5b797b7559dddfc757518875cbc8